Add more color to the webpage
I made block one the same color as the header. I don't know if i
was supposed to separate the header and block one. But I did it
anyways. so lets see how it turns out.

I am sure that if I am not supposed to do that I will probably
run into an issue later with it and you will hear about it in a future
commit.