Add space in the links list
I played  with quite a few things, but I realized the reason I could
not add space between the links is because they were all in the same
box. so anything I applied, affected all of them. Not each one
individually.

I added a style column thing, in CSS that only applies to <li>

Yes, I understand that this will probably cause issues down the line
but for this exercise. it should be just fine for now.