Add and Stylize copyright footer
I made a little copyright footer and vaguely made it
match the rest of the webpage.

There is still a ton of things I can do to this project.
But I believe it is better to have something finished
instead of having something perfect. If I want to do more
work on it later I can. I feel it is more important
to move on in my course work.

Thank you for the fun project @theodinproject